requirements.txt
playwright==1.58.0
pytest==9.0.2
pytest-playwright==0.7.2
allure-pytest==2.15.3
python-dotenv==1.2.2
